First fully working profile version
Graphics still look like crap, nothing is shown on the map yet and
the interaction is bad, but, we can show a profile that corresponds
to reality.

Look at the sample. To activate profile mode, click the button. Once
the dialog is visible, you can add points by left-clicking the map.
You can start profile generation with double click on the map.