Adding tracking of current point
Tracking is done both on the profile graphics directly as well as on the map. Note
however, that the tracking on the map is pseudo only (when panning or zooming the
map, the tracking info will not follow). This can be improved once the vector
layer of ol3 is fully functional. When this is done, then the tracking information
will be a map vector feature (right now, it's only a simple html element somehow
placed...)