First simple working Accordeon
This is just to demonstrate how easy it can be done. It also demonstrates the
different possibilites of closure ui. We are combining Ext like render techniques
(creation of DOM on the fly, similare to Ext) with decorating appraoch (the
goog.ui.zippy class is a pure decorator class working on existing html elements).

Content and styling of accordeon is of course only demonstration, but it should be
possible to put anything as content of the different accordeon tabs.